Split project into two modules allowing for HTTP to be imported into other projects
