+++ Package.swift
@@ -1,5 +1,17 @@
 import PackageDescription
 
 let package = Package(
-    name: "swift-http"
+    name: "swift-http",
+    targets: [
+      Target(
+        name: "example",
+        dependencies: [
+          .Target(name: "http")
+        ]
+      ),
+      
+      Target(
+        name: "http"
+      )
+    ]
 )
